requirements.txt
fastapi
uvicorn
celery
redis
# psycopg2-binary
google-generativeai
python-dotenv
langchain=0.110.3
langchain-google-genai